+++ src/Rocketeer/Services/ReleasesManager.php
@@ -88,7 +88,6 @@ class ReleasesManager
             rsort($releases);
 
             $this->releases = (array) $releases;
-            $this->releases = array_map('intval', $this->releases);
         }
 
         return $this->releases;