+++ src/Rocketeer/Tasks/Rollback.php
@@ -76,6 +76,6 @@ class Rollback extends AbstractTask
             $release = $this->releasesManager->getPreviousRelease();
         }
 
-        return (int) $release;
+        return (string) $release;
     }
 }