+++ tests/Traits/BashModules/CoreTest.php
@@ -19,7 +19,7 @@ class CoreTest extends RocketeerTestCase
     {
         $contents = $this->task->runRaw('ls', true, true);
 
-        $this->assertNumberFiles($contents);
+        $this->assertListDirectory($contents);
     }
 
     public function testCanCheckStatusOfACommand()
@@ -89,7 +89,7 @@ class CoreTest extends RocketeerTestCase
         $this->task->setLocal(true);
         $contents = $this->task->runRaw('ls', true, true);
 
-        $this->assertNumberFiles($contents);
+        $this->assertListDirectory($contents);
     }
 
     public function testCanConvertDirectorySeparators()