+++ src/config/remote.php
@@ -44,6 +44,14 @@ return [
     // An array of commands to run under shell
     'shelled'        => ['which', 'ruby', 'npm', 'bower', 'bundle', 'grunt'],
 
+    // Enable use of sudo for some commands
+    // You can specify a sudo user by doing
+    // 'sudo' => 'the_user'
+    'sudo' => false,
+
+    // An array of commands to run under sudo
+    'sudoed' => [],
+
     // Permissions$
     ////////////////////////////////////////////////////////////////////